requirements.txt
fastapi[standard]
keras
pillow
numpy
tensorflow
